update-ms-repo
Microsoft change the way dotnetcore installation on ubuntu thus this is the new
method. The old one we have is still active but no new package updated.

Source of truth is from this site:
https://www.microsoft.com/net/download/linux-package-manager/ubuntu16-04/sdk-current

or https://www.microsoft.com/net/download/linux which contains all other links
to lead to the above url.

Also in this change add a test-role.yml playbook to test the role.

The test-role.yml run has passed. dotnetcore sdk is installed into the container
and all others packages such as mono and nuget etc.